fix: create public-api
